Add the serialized rule as a label for the snapshot
